Intel Image Classification Flask App/Intel Image Flask App/requirements.txt
Flask==1.1.1
numpy==1.19.3
keras==2.4.3
Pillow==8.1.0
gunicorn